Feature: daemons: Use dbus to detect systemd resource startup.
The executor just has to register a callback that gets called when
JobRemoved signals are received from systemd.  This callback then takes
the data it's given and finds a matching svc_action_t.  If there's a
match, it can then finalize the action and remove it from the table of
actions we're watching for.

At the moment this only handles start actions, which is not sufficient
to close T25, but it's a good start and should be easy enough to build
on to handle stop actions and perhaps also monitors.

Related T25