Low: libpacemaker: Don't assert on injecting nonexistent nodes.
If crm_simulate tries to inject a node up or node fail action for a
nonexistent node, it'll just assert.  Instead, log an error message and
return NULL.

That NULL then needs to be propagated up a couple layers to functions
that actually expect and know how to handle errors.

Fixes T945