Low: python: Use coredumpctl to check for core files.
If we're running on a system that supports systemd, core dumps will be
managed using coredumpctl.  We need to use that tool to figure out if
there are any core dumps instead of looking in various directories.

Unfortunately, this also means we won't know whether a pacemaker-related
tool wrote a core dump or not.  They'll all be in the journal by binary
name, which means we would have to search for every pacemaker or
corosync binary possible.

However, it seems likely that ctslab is being run on a dedicated system
and so if there are any core dumps present, they likely came from
pacemaker or something it relies upon.  So it's probably fine to just
check for any core dumps and warn if one is found.

Fixes T651