Refactor: doc: Don't use autosummary in the python documentation.
Before, I was using autosummary to generate the docs for us, but as the
previous commit explained, this caused various problems.  In addition,
it's really just not necessary - we only have a couple python modules,
we're adding them at an extremely slow rate, and we're probably never
going to have more than a handful anyway.  Adding docs by hand is just
not that much of a burden.

In addition, conditionalize the index to not show up when we are
generating the singlehtml format.  It doesn't work there, so we
shouldn't add it.