Refactor: libpacemaker: Create scheduler object in pcmk__profile_dir()
The only caller is crm_simulate. This makes the profile_file() helper
easier to reason about, by guaranteeing certain properties of the
scheduler object. (It was created in pcmk__profile_dir(), the sole
caller of profile_file(), so it's easy to see that it has no relevant
state except two potential flags.)

Note that we're now able to drop the scheduler flag setting from
crm_simulate. The only other function that uses this scheduler object is
pcmk__simulate(). The first thing it does is set scheduler flags based
on the pcmk_sim_flags (via the reset() helper), making the sets in
crm_simulate redundant.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>