Refactor: libpacemaker: Use scandir() filter in pcmk__profile_dir()
It would be more efficient to compare characters directly for "." and
".." instead of making string comparison function calls. I think
readability is more important. That overhead is insignificant compared
to the simulations that we're profiling.

We could simplify the filter further if we don't add the crm_trace
messages. I have no preference. I mostly added them as documentation so
that what we're doing is obvious. Profiling is really just for
developers.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>