Refactor: tools: update_dataset() passes nulls to pcmk__schedule_actions
Previously we were passing scheduler->input as the CIB argument, which
should be a no-op. And we were already explicitly setting the
pcmk__sched_no_counts flag.

pcmk__schedule_actions() has a use-after-free bug if its cib argument is
the same as the current scheduler->input (or if it's an ancestor or
descendant). This is another step toward dropping that argument.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>