Feature: libpacemaker: Reset scheduler object in pcmk_simulate()
pcmk_simulate() resets or overwrites most of the passed-in scheduler
object already. It resets everything if there are any injections. It
doesn't make any sense to pass in an already-initialized scheduler
object and rely on any particular behavior.

With that said, behavior may change in certain corner cases if the
scheduler argument already has some values (such as flags) set.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>