Refactor: cts: Clarify spacing in cts-exec.in.
The self._action_timeout string had a space at the beginning and end,
which is why we could get away with not having spaces around it in
format strings later.  I think this is less clear than removing the
spaces from the variable and adding them to the format strings where
appropriate.