Fix: libraries: Update version numbers in so names
This fixes a longstanding regression, introduced in dad2411 for
libstonithd and in e014c52 for libcib.

The libstonithd version mismatch currently causes issues with
dlm_controld, impacting gfs2 filesystems.

It's not clear whether anything uses the CIB_LIBRARY constant. It looks
as if it could be deprecated. Nothing internal uses it. However, dlm
includes stonith-ng.h, which uses STONITH_LIBRARY in inline functions.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>