Refactor: cts: Disable the too-many-locals warning in cts-scheduler.
Fixing this is a larger problem than is addressable in this patch set -
it requires rethinking how cts-scheduler is put together, which I don't
want to work on at the moment.  In the interest of running pylint and
pyflake as part of the CI process without failing every build, I'm
disabling this warning for now.