Build: cts: Add cts-cli to the python_files list.
This is a list of generated python scripts that can have pylint and
pyflake run against them.  It exists because we were working on these
scripts one at a time, and I didn't want to overwhelm the results with
tons of errors from scripts we hadn't gotten around to yet.

The problem with this is that sometimes, we forget to add scripts to the
list.  That's the case with cts-cli, which is why it still needs to be
converted to using f-strings.

When we convert cts-schemas to python, this can all go away and we can
use test_SCRIPTS instead.