Refactor: cts: Replace {shadow} strings in cts-cli where possible.
The string substitutions I was doing with s.format() through the
apply_substitutions function won't work once we move to f-strings.  I
don't think that function is doing anything special, however, so we can
just get rid one substitution it's performing at a time.

This removes all uses of {shadow} that don't occur in a string with some
other substitution.