Refactor: cts: Use {cts_cli_home} as an f-string where possible.
Just like the previous commit, there's still places where we can't
completely switch over to f-strings for performing this substition, so
it has to stay in the apply_substitutions function for now.

However, there are several places we no longer need to call that
function because the only thing we were substituting was cts_cli_home.

Additionally, test_home wasn't be substituted anywhere so it can go.