Refactor: cts: Get rid of the make_test_group function in cts-cli.
The idea here was to reduce test duplication by introducing a way to
create the same test for multiple classes, substituting in some piece
of data for each class.  At the moment, the only thing being substituted
in is the format.

However, this interferes with f-strings.  These are interpreted early so
whatever variable is referenced in the string has to already exist, and
you can't (or at least aren't supposed to) mix f-strings with .format().

So if we are to shift to using f-strings, all these lines have to
change.  After considering a couple approaches, I've decided to just go
with the simplest thing:  get rid of make_test_group and just duplicate
all the tests.

It's a lot of code, but I don't think it matters all that much.  The
result is simpler to follow, at least, because there's not another
TestGroup in the middle, nor is there this function call with a layer of
string substitutions.