Refactor: libcrmcommon: Remove pcmk__ipc_buffer_size.
The IPC buffer is now always 128k, so we can just have
crm_ipc_default_buffer_size return that wherever we need to know the
buffer size.  The idea here is that if a message doesn't fit in that
amount of space, we're going to split it up instead of refuse to process
it.