Refactor: libcrmcommon: Don't compress messages in pcmk__ipc_prepare_iov.
If a message is too big to fit in a single buffer's worth, we are
compressing it.  However, we still have a size limit - if the compressed
message is too big to fit, there's nothing else we can do except error.

We are moving towards splitting up too large messages into multiple
smaller ones that do fit.  Once we start doing this, there's not much
point in compressing the message.  I think we're just adding compression
and decompression time to message processing for little benefit.

With this commit, servers will no longer compress IPC messages.  This
will cause problems with too-large messages until splitting is added.