Refactor: libcrmcommon: pcmk__ipc_prepare_iov takes a string...
...instead of an xmlNode.  The idea here is that to deal with split up
messages, this function can be called repeatedly from a loop.  Each call
will prepare the next chunk of the message to be sent.  The first step
in doing that is simply changing the type of the parameter.  This does
result in a little code duplication in the callers, but there are only a
couple.