Refactor: libcrmcommon: pcmk__ipc_prepare_iov should take an offset.
The function now only prepares as much for one I/O vector as will fit in
a single buffer.

The idea is that you can call the function repeatedly in a loop,
preparing and transmitting individual chunks of a single XML message.
As long as pcmk__ipc_prepare_iov returns EAGAIN, you know there's more
that needs to be prepared.  Keep a running count of the bytes it
prepared and pass that in for offset.

Note that clients are not doing this yet, so very large IPC messages are
still going to fail.