Refactor: libcrmcommon: Reduce IPC client receive code duplication.
Right now, there are two pathways in crm_ipc_send for receiving a reply
to an IPC request - one if there's a timeout, and one if there is not.
This also means that only the timeout path tries to decompress the reply
and checks if it's receiving replies that are too old or too new.  It
also means that the reply can be converted into XML more than one time,
since both crm_ipc_send and its internal helper will do that.

Instead, make sure both pathways call into internal_ipc_get_reply,
setting up the timeout parameters appropriately.