Low: tools: Dry-run crm_resource --fail if CIB_file is set
We're already doing this for --cleanup and --refresh.

If CIB_file is set, our CIB connection will be to the file instead of to
the live CIB. It's nonsensical or dangerous to try to connect to the
controller and fail a resource that we found in a CIB file. So I
consider the existing behavior a bug.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>