Refactor: fencer: Fix coverity suppression for rand()
The suppression needs to use event tag dont_call rather than dontcall.
It also needs to be on the line immediately before the rand() call.

This commit also simplifies the block in which the rand() call occurs.
And it adds 1 to the modulus so that we include the full range of
(delay_max - delay_base), not that it really matters.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>