Refactor: libcrmcommon: New pcmk__xpath_result()
To replace getXpathResult().

We also create a pcmk__xpath_match_element() function to preserve
behavior of public API functions and of ACLs.

The pcmk__xpath_match_element() call documented in cib_process_xpath()
will be removed in an upcoming commit. We needed to introduce
pcmk__xpath_result() before we could fix it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>