Refactor: libcib: Drop dedupXpathResults() internally
This addresses what ca271b6 addressed, but without the need to
deduplicate results.

Modify and replace operations still have the potential for
use-after-free since we continue to iterate in forward order. However,
we don't make those calls with cib_multiple internally. External callers
might rely on the forward processing somehow... though it's not clear if
that's really possible without a segfault.

Anyway, in the past we deduplicated results only for delete operations,
and this commit preserves that behavior.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>