Refactor: libcrmcommon: New pcmk__xpath_foreach_result()
To replace crm_foreach_xpath_result().

We can add a reverse argument later if needed.

Note that all internal callers are passing XPath expressions that can
only match elements. So we don't have to worry about the "return an
element for an attribute, etc." logic that we were using in the past via
crm_foreach_xpath_result() -> pcmk__xpath_match_element() (or
getXpathResult() farther in the past).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>