Refactor: libcrmcommon: Drop pcmk__tracking_xml_changes()
Most of these changes are straightforward because all but one caller
passed false as the lazy argument. This is equivalent to checking the
the pcmk__xf_tracking flag on the node's doc.

One caller passes true, which also checks the pcmk__xf_lazy flag.

It's probably worth renaming pcmk__xf_lazy to something with clearer
meaning later.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>