Refactor: libcrmcommon: Take doc as argument in pcmk__set_xml_doc_flag()
IMO this makes it clearer that we're setting flags for a particular
document. It has nothing to do with any particular node within the
document; we were using the node argument only to access its doc member.

Also add doxygen, rename to pcmk__xml_doc_set_flags(), and take a
uint32_t argument instead of an enum (since it may be a group of
multiple flags).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>