Refactor: libcib, libcrmcommon: Drop xml_track_changes() internally
The only remaining uses were in cib_perform_op() and in the unit tests
for pcmk__xc_create().

For cib_perform_op(), this introduces some duplication. We could
functionize the duplicated code, but that entire function is a huge mess
anyway, so I'm simply leaving it for now until we can revisit it as a
whole. Note that pcmk__enable_acl() is equivalent to the three calls
that xml_track_changes() was making.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>