Refactor: libstonithd: Query stonith devices more clearly
Previously, we performed an XPath search that gets all "agent"
attributes. But for XPath search results that are not element nodes,
getXpathResult() an element (or NULL). In the case of an attribute, it
returns the attribute's parent element.

This is not at all clear. So instead, we now directly search for
elements that possess an "agent" attribute.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>