Refactor: libcrmcommon: Rename xml_private_flags to pcmk__xml_flags
And add doxygen. The doxygen is a bit hand-wavy for a few of them,
because I don't yet understand everything they're doing.

The wide spacing before the equal sign is to allow room to do a planned
rename of pcmk__xf_lazy later, without having to change the other lines
again.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>