Refactor: libcrmcommon: Don't set XML private flags without tracking
Previously, we marked newly a created element, attribute, or comment as
"dirty and created" and marked all of its parents as dirty, regardless
of whether tracking was enabled for the document. Similarly, we marked a
newly created document as "dirty and created," even though tracking
could not have been enabled for it.

This didn't make sense, and it often required committing changes that
were made without tracking enabled, to ensure that those changes didn't
show up in a patchset after tracking was enabled later.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>