Refactor: libcrmcommon: New pcmk__xml_mark_changes()
To replace xml_calculate_changes().

This is just the existing mark_xml_changes(), except that it sets
pcmk__xf_tracking on the document for convenience if it's not already
set. This only needs to be done for the top-level call, but there's no
harm in doing it on the recursive calls; it just adds redundant trace
logging.

The new CRM_CHECK() in xml_calculate_significant_changes() is just what
would have happened in its call to xml_calculate_changes() prior to this
commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>