Refactor: tools: Clean up crm_diff.c:apply_patchset()
We don't do anything with input afterward, and this is static for a CLI
tool, so no need to make a copy.

Don't bother adding doxygen. We're about to drop this function. I'm
doing it in two steps to clarify what it's doing and that it's no longer
worth functionizing.

Also NULL-check digest before using it in crm_strdup_printf() in
pcmk__digest_xml().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>