Refactor: tools: Clean up crm_diff.c:apply_patch()
We don't do anything with input afterward, and this is static for a CLI
tool, so no need to make a copy.

Don't calculate a digest. It's only for tracing, and it's unclear what
we would use it for. The commit that added it (2cff40d2) doesn't give
any explanation.

Don't bother adding doxygen. We're about to drop this function. I'm
doing it in two steps to clarify what it's doing and that it's no longer
worth functionizing.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>