Low: tools: Be stricter about crm_diff --cib/--no-version
Previously, we allowed --cib and --no-version to be specified together
as long as --patch was also specified. In that case, --no-version was
ignored.

Our option validation (or lack thereof) in crm_diff is a big mess. This
is one small improvement that SHOULDN'T break anyone's workflow... if it
does, they can fix it, or if it's something really insane, we can
revert.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>