Feature: tools: Use formatted output in crm_diff.c
It's a shame that the pcmk__output_t:output_xml() method requires
preformatting the XML into a string. Since it's an internal API, maybe
we can revisit that soon.

Also, we already have a diff schema in xml/api. We **could** use it and
output the patchset XML directly, instead of outputting the patchset XML
as CDATA. Ken was pretty adamant about outputting CDATA here. I don't
know why, or can't remember. The argument can't be that the XML isn't
generated by crm_diff itself, because that's true for almost all of our
output formatters. Something about how the format is supposed to be
meaningful only to Pacemaker... though I don't see why that ought to be
the case.

I'm fine with CDATA, regardless.

Ref T114

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>