Fix: libcrmcommon: provide a description for pcmk_rc_no_dc return code
This adds an info entry for `pcmk_rc_no_dc` return code so that
pcmk_rc_str() provides an informative description for it rather than
just an "Error".

This should have been done with a1d94f7ab5.

In particular for the output of `crmadmin --dc_lookup/-D`, dc() function
outputs somewhat redundant information when DC is not yet elected. But
I'm not sure whether/how we should change the output format there.