Refactor: various: Check pcmk__xe_get_time() rc only where useful
Take advantage of the fact that the destination output variable is
unchanged on error. If the caller initialized it to some value (either
explicitly or by allocating memory with calloc() or
pcmk__assert_alloc()), then we can assume that either it's unchanged or
a valid value was parsed into it.

The lrmd_dispatch_internal() case is a little bit different. We weren't
checking the return value. epoch would've been unchanged from
PCMK__XA_LRMD_RUN_TIME if PCMK__XA_LRMD_RCCHANGE_TIME were unset. But it
won't be unset (see send_cmd_complete_notify()), so this is irrelevant,
and the epoch variable is unnecessary.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>