Refactor: libcrmcommon: New pcmk__xe_get_guint()
To replace crm_element_value_ms().

The only thing milliseconds-specific about the existing function was the
error message. IMO, it's better not to log errors or warnings within the
getter function. The caller has more context and can log errors or
warnings if appropriate. Sometimes, a parse failure may be expected as a
legitimate possibility.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>