Refactor: libcrmcommon: New pcmk__xe_set_ll()
The only reason we return a value here is so that pcmk__xe_set_timeval()
can check it. We could have a static helper function that both of them
call, so that this will be consistent, but I can't be bothered.

The existing crm_xml_add_ll() prints into a static buffer. While that is
undeniably more efficient, I don't think it matters.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>