Refactor: libcrmcommon: Allow negative values in pcmk__parse_ms()
All callers except crm_mon already handled this.

Also, crm_mon was the only caller that passed NULL for the result
argument. So we now CRM_CHECK() that it's not NULL. It's easty add "NULL
result" as a feature again in the future, if there's a frequent need to
validate that a time+units string is parsable without actually needing
its value.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>